% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_DRpower_loaded}
\alias{check_DRpower_loaded}
\title{Check that DRpower package has loaded successfully}
\usage{
check_DRpower_loaded()
}
\description{
Simple function to check that DRpower package has loaded 
  successfully. Prints "DRpower loaded successfully!" if so.
}
