% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_sim}
\alias{df_sim}
\title{Summary of simulations from the threshold analysis}
\format{
A data.frame of 547200 rows and 19 columns. The first 13 columns give
  parameter combinations that were used in simulating and analysing data. The
  "reps" column gives the number of times simulation was repeated, and "seed"
  gives the value of the seed that was used at the start of this loop (to
  ensure reproducibility). "prev_thresh" gives the prevalence threshold used
  in hypothesis testing. The final three columns give the estimates power
  over simulations along with lower and upper 95\% CIs calculated using the
  method of Clopper and Pearson (1934).
}
\usage{
data(df_sim)
}
\description{
This object was produced by running the function
  \code{get_power_threshold()} over a wide range of parameter combinations.
  This data.frame contains the results of these simulations attached to the
  parameter values used in simulation. The most obvious use of this object is
  in constructing power curves over a range of sample sizes (see
  \code{?power_curve()}).
}
\examples{
data(df_sim)

}
\keyword{datasets}
