% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_ss}
\alias{df_ss}
\title{Minimum sample sizes for the threshold analysis}
\format{
A data frame of 6840 rows and 15 columns. The first 14 columns give
  parameter combinations that were used in simulating and analysing data. The
  final "N_opt" column gives the optimal sample size to achieve a power of
  80%.
}
\usage{
data(df_ss)
}
\description{
TODO
}
\examples{
data(df_ss)

}
\keyword{datasets}
