% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRpower.R
\docType{package}
\name{DRpower}
\alias{DRpower-package}
\alias{DRpower}
\title{The DRpower app for design and analysis of Plasmodium falciparum
pfhrp2/3 data}
\description{
This package can be used in the design and/or analysis stages of
  Plasmodium falciparum pfhrp2/3 deletion prevalence studies. We assume that
  the study takes the form of a clustered prevalence survey, meaning the data
  consists of a numerator (number tested) and denominator (number of
  deletions found) over multiple clusters. We are interested in estimating
  the study-level prevalence, i.e. over all clusters, while accounting for
  the possibility of high intra-cluster correlation. The analysis approach
  uses a Bayesian random effects model to estimate prevalence and
  intra-cluster correlation. The approach to power analysis is
  simulation-based, running the analysis many times on simulated data and
  estimating empirical power. This method can be used to establish a minimum
  sample size required to achieve a given target power.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/mrc-ide/DRpower/issues}
}

}
\author{
\strong{Maintainer}: Bob Verity \email{r.verity@imperial.ac.uk}

Authors:
\itemize{
  \item Shazia Ruybal \email{s.ruybal@imperial.ac.uk}
}

}
