% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{estimate_power}
\alias{estimate_power}
\title{Estimate power via simulation}
\usage{
estimate_power(
  clusters,
  total_samples,
  prevalence,
  ICC = NULL,
  prevalence_threshold = 0.05,
  alpha = 0.05,
  reps = 100,
  fix_ICC_estimate = NULL
)
}
\arguments{
\item{clusters}{number of clusters.}

\item{total_samples}{total sample size per cluster.}

\item{prevalence}{assumed true prevalence of pfhrp2 deletions. Input as
proportion between 0 and 1.}

\item{ICC}{assumed true intra-cluster correlation (ICC), between 0 and 1.}

\item{prevalence_threshold}{threshold used in decision-making. Input as
proportion between 0 and 1.}

\item{alpha}{the significance level of the credible interval - for example,
use \code{alpha = 0.05} for a 95\% interval.}

\item{reps}{number of times to repeat simulation per parameter combination.}

\item{fix_ICC_estimate}{if defined, fix the estimated ICC at this value
rather than estimating from data. Note, this does not change the ICC used
in simulation, which is still taken from \code{ICC}.}
}
\description{
Estimates power empirically via repeated simulation. A range of
  input parameters can be specified (e.g. a range of sample sizes per
  cluster), power is then calculated on all combinations of input parameters.
}
\details{
Estimates power using the following approach:
\enumerate{
 \item Simulate repeatedly from the function \code{rbbinom_reparam()}
 using known values (e.g. a known "true" prevalence).
 \item Construct confidence intervals (CIs) on the prevalence from simualted
 data using \code{estimate_prevalence}.
 \item Make a decision as to whether prevalence is above the threshold,
 below, or inconclusive based on CIs.
 \item Count the number of simulations for which the correct conclusion is
 reached. This gives an estimate of empirical power, and upper and lower 95%
 binomial CIs on the power are produced using the method of Clopper and
 Pearson (1934).
}
}
\references{
Clopper, C.J. and Pearson, E.S., 1934. The use of confidence or fiducial
limits illustrated in the case of the binomial. Biometrika, 26, 404–413. doi:
10.2307/2331986.
}
