% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{estimate_prevalence}
\alias{estimate_prevalence}
\title{Estimate prevalence of pfhrp2 deletions from clustered data}
\usage{
estimate_prevalence(
  pos_samples,
  total_samples,
  alpha = 0.05,
  fix_ICC_estimate = NULL
)
}
\arguments{
\item{pos_samples}{number of "positive" samples per cluster.}

\item{total_samples}{total sample size per cluster.}

\item{alpha}{the significance level of the credible interval - for example,
use \code{alpha = 0.05} for a 95\% interval.}

\item{fix_ICC_estimate}{if defined, fix the ICC at this value rather than estimating
from data.}
}
\description{
Takes raw counts of pfhrp2 deletions in multiple clusters (e.g.
  clinics) and estimates the overall prevalence. Uses a standard analysis
  approach in which the intra-cluster correlation (ICC) is first estimated,
  and this is used to compute an "effective sample size". Binomial confidence
  intervals are then calculated using this effective sample size rather than
  the raw sample size.
}
\details{
The ICC is estimated via the ICCbin package using the stabilised
  estimate (\code{stab}) method proposed by Tamura and Young (1987). The
  design effect, \eqn{D_{eff}}{Deff}, is then calculated via the formula:
  \deqn{D_{eff} = 1 + (m - 1)r}{Deff = 1 + (m - 1)*r}
  where \eqn{m} is the number of clusters and \eqn{r} is the ICC. The
  effective sample size, \eqn{m_{eff}}{meff} is calculated via the formula:
  \deqn{m_{eff} = \frac{n m}{D_{eff}}}{meff = n*m / Deff}
  Finally, binomial confidence intervals are calculated using the
  Clopper-Pearson (1934) interval at a significance level \code{alpha}
  provided by the user (two-tailed).
  
  In some cases it is not possible to estimate the ICC using the stabilised
  method above - for example when there is a single cluster or when all
  counts are zero. In these cases a default value of 0.05 is used, which
  corresponds to a design effect of 2.8 for the reference situation of 37
  samples per clinic as described in the WHO master protocol.
}
\references{
Tamura, R.N. and Young, S.S., 1987. A stabilized moment estimator for the
beta-binomial distribution. Biometrics, pp.813-824.

Clopper, C.J. and Pearson, E.S., 1934. The use of confidence or fiducial
limits illustrated in the case of the binomial. Biometrika, 26, 404–413. doi:
10.2307/2331986.
}
