% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_credible_ICC}
\alias{get_credible_ICC}
\title{Get credible intervals for the intra-cluster correlation coefficient}
\usage{
get_credible_ICC(k, m, alpha = 0.05, p_nodes = 20, rho_nodes = 20)
}
\arguments{
\item{k}{counts of positive samples.}

\item{m}{sample size per cluster.}

\item{alpha}{the significance level of the credible interval - for sample,
use \code{alpha = 0.05} for 95}

\item{p_nodes, rho_nodes}{number of nodes to use in the Gaussian quadrature
approximate integral method.}
}
\description{
Get lower and upper credible intervals on the intra-cluster
  correlation coefficient (ICC).
}
\examples{
# TODO
print("foo")
}
