% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes.R
\name{get_credible_prevalence_fast}
\alias{get_credible_prevalence_fast}
\title{Rcpp implementation of \code{get_credible_prevalence()}}
\usage{
get_credible_prevalence_fast(
  n,
  N,
  alpha = 0.05,
  prior_prev_shape1 = 1,
  prior_prev_shape2 = 1,
  prior_ICC_shape1 = 1,
  prior_ICC_shape2 = 1
)
}
\arguments{
\item{n, N}{the numerator (\code{n}) and denominator (\code{N}) per cluster.}

\item{alpha}{the significance level of the credible interval - for example,
use \code{alpha = 0.05} for a 95\% interval.}

\item{prior_prev_shape1, prior_prev_shape2, prior_ICC_shape1, prior_ICC_shape2}{parameters that dictate the shape of the priors on prevalence and the ICC.
Increasing the first shape parameter (e.g. \code{prior_p_shape1}) pushes
the distribution towards 1, increasing the second shape parameter (e.g.
\code{prior_p_shape2}) pushes the distribution towards 0. Increasing both
shape parameters squeezes the distribution and makes it narrower.}
}
\description{
Equivalent to \code{get_credible_prevalence()} in functionality,
  but implemented in Rcpp for speed.
}
