% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_joint_grid}
\alias{get_joint_grid}
\title{Get posterior distribution of both prevalence and the ICC on a grid}
\usage{
get_joint_grid(
  n,
  N,
  prior_prev_shape1 = 1,
  prior_prev_shape2 = 1,
  prior_ICC_shape1 = 1,
  prior_ICC_shape2 = 1,
  prev_cells = 64,
  ICC_cells = 64
)
}
\arguments{
\item{n, N}{the numerator (\code{n}) and denominator (\code{N}) per cluster.}

\item{prior_prev_shape1, prior_prev_shape2, prior_ICC_shape1, prior_ICC_shape2}{parameters that dictate the shape of the priors on prevalence and the ICC.
Increasing the first shape parameter (e.g. \code{prior_p_shape1}) pushes
the distribution towards 1, increasing the second shape parameter (e.g.
\code{prior_p_shape2}) pushes the distribution towards 0. Increasing both
shape parameters squeezes the distribution and makes it narrower.}

\item{prev_cells, ICC_cells}{the number of cells in the grid in each
dimension.}
}
\description{
Get posterior distribution of both prevalence and the ICC on a
  grid.
}
