% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_quadrature.R
\name{get_max_x}
\alias{get_max_x}
\title{Get maximum of interpolated curve}
\usage{
get_max_x(df_norm)
}
\arguments{
\item{df_norm}{a normalised quadrature data.frame as produced by
\code{normalise_quadrature()}.}
}
\description{
Returns the x-value at which y reaches its maximum value under
  an adaptive quadrature object.
}
