% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_power}
\alias{get_power}
\title{Estimate power via simulation}
\usage{
get_power(
  cluster_size,
  prevalence,
  ICC,
  prevalence_threshold = 0.05,
  alpha = 0.05,
  reps = 100
)
}
\arguments{
\item{cluster_size}{vector giving the number of samples obtained from each
cluster.}

\item{prevalence}{assumed true prevalence of pfhrp2 deletions. Input as
proportion between 0 and 1.}

\item{ICC}{assumed true intra-cluster correlation (ICC), between 0 and 1.}

\item{prevalence_threshold}{threshold used in decision-making. Input as
proportion between 0 and 1.}

\item{alpha}{the significance level of the credible interval - for example,
use \code{alpha = 0.05} for a 95\% interval.}

\item{reps}{number of times to repeat simulation per parameter combination.}
}
\description{
Estimates power empirically via repeated simulation. A range of
  input parameters can be specified (e.g. a range of sample sizes per
  cluster), power is then calculated on all combinations of input parameters.
}
\details{
Estimates power using the following approach:
\enumerate{
 \item Simulate repeatedly from the function \code{rbbinom_reparam()} using
 known values (e.g. a known "true" prevalence and intra-cluster correlation).
 \item Construct credible intervals (CrIs) on the prevalence from simulated
 data using \code{get_credible_prevalence()}.
 \item Make a decision as to whether prevalence is above the threshold,
 below, or inconclusive based on CrIs.
 \item Count the number of simulations for which the correct conclusion is
 reached. This gives an estimate of empirical power, and upper and lower 95%
 binomial CIs on the power are produced using the method of Clopper and
 Pearson (1934).
}
}
\references{
Clopper, C.J. and Pearson, E.S., 1934. The use of confidence or fiducial
limits illustrated in the case of the binomial. Biometrika, 26, 404–413. doi:
10.2307/2331986.
}
