% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_joint}
\alias{plot_joint}
\title{Contour plot of joint posterior distribution of prevalence and ICC}
\usage{
plot_joint(
  n,
  N,
  prior_prev_shape1 = 1,
  prior_prev_shape2 = 1,
  prior_ICC_shape1 = 1,
  prior_ICC_shape2 = 9,
  prev_breaks = seq(0, 1, 0.01),
  ICC_breaks = seq(0, 1, 0.01),
  n_bins = 5
)
}
\arguments{
\item{n, N}{the numerator (\code{n}) and denominator (\code{N}) per cluster.
These are both integer vectors.}

\item{prior_prev_shape1, prior_prev_shape2, prior_ICC_shape1, prior_ICC_shape2}{parameters that dictate the shape of the Beta priors on prevalence and the
ICC. See the \href{https://en.wikipedia.org/wiki/Beta_distribution}{Wikipedia page on the Beta
distribution} for more detail. The default values of these parameters were chosen
based on an
\href{https://mrc-ide.github.io/DRpower/articles/historical_analysis.html}{analysis
of historical pfhrp2/3 studies}, although this does not guarantee that they
will be suitable in all settings.}

\item{prev_breaks, ICC_breaks}{the values at which to evaluate the posterior
in both dimensions. Prevalence is returned in columns, ICC in rows.}

\item{n_bins}{the number of equally spaced breaks in the contour plot. For
example, 5 bins creates 4 contour lines at 20\%, 40\%, 60\% and 80\% of
the maximum value.}
}
\description{
Runs \code{get_joint()} to obtain the joint posterior
  distribution of the prevalence and the ICC. Creates a ggplot contour plot
  object from this result.
}
\examples{
plot_joint(n = c(5, 2, 9), N = c(100, 80, 120))

}
