% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior}
\alias{plot_posterior}
\alias{plot_prevalence}
\alias{plot_ICC}
\title{Plot posterior distribution of prevalence and ICC}
\usage{
plot_prevalence(
  n,
  N,
  prev_range = c(0, 1),
  alpha = 0.05,
  prev_thresh = 0.05,
  prior_prev_shape1 = 1,
  prior_prev_shape2 = 1,
  prior_ICC_shape1 = 1,
  prior_ICC_shape2 = 9,
  site_weights = rep(1, length(n)),
  CrI_type = "HDI",
  n_intervals = 20,
  use_cpp = TRUE
)

plot_ICC(
  n,
  N,
  ICC_range = c(0, 1),
  alpha = 0.05,
  prev_thresh = 0.05,
  prior_prev_shape1 = 1,
  prior_prev_shape2 = 1,
  prior_ICC_shape1 = 1,
  prior_ICC_shape2 = 9,
  site_weights = rep(1, length(n)),
  CrI_type = "HDI",
  n_intervals = 20,
  use_cpp = TRUE
)
}
\arguments{
\item{n, N}{the numerator (\code{n}) and denominator (\code{N}) per cluster.
These are both integer vectors.}

\item{prev_range}{the range of prevalence values explored. Vector of two
values giving lower and upper limits, defined between 0 and 1.}

\item{alpha}{the significance level of the credible interval - for example,
use \code{alpha = 0.05} for a 95\% interval. See also \code{CrI_type}
argument for how this is calculated.}

\item{prev_thresh}{the prevalence threshold that we are testing against
(single value only, proportion between 0 and 1).}

\item{prior_prev_shape1, prior_prev_shape2, prior_ICC_shape1, prior_ICC_shape2}{parameters that dictate the shape of the Beta priors on prevalence and the
ICC. See the \href{https://en.wikipedia.org/wiki/Beta_distribution}{Wikipedia page on the Beta
distribution} for more detail. The default values of these parameters were chosen
based on an
\href{https://mrc-ide.github.io/DRpower/articles/historical_analysis.html}{analysis
of historical pfhrp2/3 studies}, although this does not guarantee that they
will be suitable in all settings.}

\item{site_weights}{optional weighting factors applied to each site. The
log-likelihood of each site is multiplied by this factor, resulting in a
pseudo-likelihood based method. This is only an approximate solution to the
problem of estimating a weighted prevalence - a better solution is to
choose sites with probabilities proportional to burden, in which case no
additional weighting factor is required.}

\item{CrI_type}{which method to use when computing credible intervals.
Options are "ETI" (equal-tailed interval) or "HDI" (high-density interval).
The ETI searches a distance \code{alpha/2} from either side of the [0,1]
interval. The HDI method returns the narrowest interval that subtends a
proportion \code{1-alpha} of the distribution. The HDI method is used by
default as it guarantees that the MAP estimate is within the credible
interval, which is not always the case for the ETI.}

\item{n_intervals}{the number of intervals used in the adaptive quadrature
method. Increasing this value gives a more accurate representation of the
true posterior, but comes at the cost of reduced speed.}

\item{use_cpp}{if \code{TRUE} (the default) then use an Rcpp implementation
of the adaptive quadrature approach that is much faster than the base R
method.}

\item{ICC_range}{the range of ICC values explored. Vector of two values
giving lower and upper limits, defined between 0 and 1.}
}
\description{
These two functions run \code{get_prevalence()} and
  \code{get_ICC()} respectively to obtain the full posterior distribution of
  the parameter of interest. Then they plot the posterior density along with
  some useful visualisations including the 95% CrI.
}
\examples{
plot_prevalence(n = c(5, 2, 9), N = c(100, 80, 120))

plot_ICC(n = c(5, 2, 9), N = c(100, 80, 120))

}
