% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_power}
\alias{plot_power}
\title{Plot a power curve using pre-computed values}
\usage{
plot_power(
  n_clust = 5,
  prevalence = 0.1,
  ICC = 0.05,
  prev_thresh = 0.05,
  N_min = 1,
  N_max = 2000
)
}
\arguments{
\item{n_clust}{the number of clusters. Allowed values are anything in
\code{2:20}, including vectors of values.}

\item{prevalence}{the assumed true prevalence of pfhrp2/3 deletions in the
domain. Allowed values are anything in \code{seq(0, 0.2, 0.01)}, including
vectors of values.}

\item{ICC}{the assumed intra-cluster correlation. Allowed values are" \{0,
0.01, 0.02, 0.05, 0.1, 0.2\}.}

\item{prev_thresh}{the prevalence threshold against which we are comparing.
Allowed values are: \{0.05, 0.08, 0.1\}.}

\item{N_min, N_max}{plotting limits on the x-axis.}
}
\description{
Runs \code{get_joint()} to obtain the joint posterior
  distribution of the prevalence and the ICC. Creates a ggplot contour plot
  object from this result.
}
\examples{
plot_power()

}
