% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_quadrature.R
\name{plot_quadrature}
\alias{plot_quadrature}
\title{Plot output of adaptive quadrature}
\usage{
plot_quadrature(f1, df_quad)
}
\arguments{
\item{f1}{the function used in adaptive quadrature. This is needed to compare
against the adaptive quadrature result by running on a grid.}

\item{df_quad}{a data.frame output from the function
\code{adaptive_quadrature()}.}
}
\description{
Takes data.frame output from the function
  \code{adaptive_quadrature()}, normalises through the function
  \code{normalise_quadrature()}, and produces a plot comparing the
  appximation with the true function evaluated over a grid.
}
