% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_quadrature.R
\name{pquad}
\alias{pquad}
\title{Integrate adaptive quadrature object up to p}
\usage{
pquad(df_norm, p = 0.5)
}
\arguments{
\item{df_norm}{a normalised quadrature data.frame as produced by
\code{normalise_quadrature()}.}

\item{p}{the target area}
}
\description{
Returns the area under the curve of an adaptive quadrature
  object up to x=p (similar to e.g. \code{pnorm()}).
}
