% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_quadrature.R
\name{qquad}
\alias{qquad}
\title{Get quantile from adaptive quadrature object}
\usage{
qquad(df_norm, q = 0.5)
}
\arguments{
\item{df_norm}{a normalised quadrature data.frame as produced by
\code{normalise_quadrature()}.}

\item{q}{the target quantile.}
}
\description{
Returns the x-value for an adaptive quadrature object at which
  the integral up to this value equals the given quantile (similar to e.g.
  \code{qnorm()}).
}
