% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes.R
\name{get_credible_ICC}
\alias{get_credible_ICC}
\title{Get credible intervals for the intra-cluster correlation coefficient
  (ICC)}
\usage{
get_credible_ICC(
  n,
  N,
  alpha = 0.05,
  prior_prev_shape1 = 1,
  prior_prev_shape2 = 1,
  prior_ICC_shape1 = 1,
  prior_ICC_shape2 = 1,
  debug_on = FALSE,
  debug_grid = 1000
)
}
\arguments{
\item{n, N}{the numerator (\code{n}) and denominator (\code{N}) per cluster.}

\item{alpha}{the significance level of the credible interval - for example,
use \code{alpha = 0.05} for a 95\% interval.}

\item{prior_prev_shape1, prior_prev_shape2, prior_ICC_shape1, prior_ICC_shape2}{parameters that dictate the shape of the priors on prevalence and the ICC.
Increasing the first shape parameter (e.g. \code{prior_p_shape1}) pushes
the distribution towards 1, increasing the second shape parameter (e.g.
\code{prior_p_shape2}) pushes the distribution towards 0. Increasing both
shape parameters squeezes the distribution and makes it narrower.}

\item{debug_on}{For use in debugging, for advanced users only. If \code{TRUE}
then produces a plot of the posterior distribution comparing various
computational methods and approximations that are used internally. The
black solid line is the distribution of prevalence marginalised over rho
via trapezoidal rule on a fine grid (see \code{debug_grid} argument). This
can be used as a sanity check, as for a fine enough grid this must tend
towards the correct answer. The dashed red line is the distribution of
prevalence marginalised using Gaussian quadrature (GQ) instead of
trapezoidal rule. In the full method, GQ is only used at a smaller number
of points indicated by green circles. The dashed green line is the
quadratic interpolation of these points via Simpson's rule. The final
credible interval estimates are calculated from the dashed green line. If
the method is working correctly we should see a dashed green and red line
that overlays a solid black line so closely that this almost looks like a
black border.}

\item{debug_grid}{The number of equally spaced intervals used when performing
marginalisation via trapezoidal rule in debugging (see \code{debug_on}
argument).}
}
\description{
Produces lower and upper credible intervals on the intra-cluster
  correlation coefficient (ICC) from clustered counts. By default these are
  95\% credible intervals, although the significance level can be altered by
  changing the \code{alpha} input value.
}
\details{
There are two unknown quantities in the DRpower model: the
  prevalence and the ICC. This function integrates out the prevalence over a
  prior distribution to give the marginal distribution of the ICC. Then it
  returns the credible intervals of this distribution at a specified
  significance level.
}
\examples{
# define three clusters with different number of observed positive counts.
# Try the default 95\% CrI and a more stringent significance level
get_credible_ICC(n = c(2, 5, 4), N = 10)
get_credible_ICC(n = c(2, 5, 4), N = 10, alpha = 0.01)
}
