% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes.R
\name{get_credible_ICC}
\alias{get_credible_ICC}
\title{Get credible intervals for the intra-cluster correlation coefficient
  (ICC)}
\usage{
get_credible_ICC(
  pos_samples,
  total_samples,
  alpha = 0.05,
  prior_prev_shape1 = 1,
  prior_prev_shape2 = 1,
  prior_ICC_shape1 = 1,
  prior_ICC_shape2 = 1
)
}
\arguments{
\item{pos_samples}{number of "positive" samples per cluster.}

\item{total_samples}{total sample size per cluster.}

\item{alpha}{the significance level of the credible interval - for example,
use \code{alpha = 0.05} for a 95\% interval.}

\item{prior_prev_shape1, prior_prev_shape2, prior_ICC_shape1, prior_ICC_shape2}{parameters that dictate the shape of the priors on prevalence and the ICC.
Increasing the first shape parameter (e.g. \code{prior_p_shape1}) pushes
the distribution to the right, increasing the second shape parameter (e.g.
\code{prior_p_shape2}) pushes the distribution to the left. Increasing both
shape parameters squeezes the distribution and makes it narrower.}
}
\description{
Produces lower and upper credible intervals on the intra-cluster
  correlation coefficient (ICC). By default these are 95\% credible
  intervals, although the user can choose any significance level by altering
  the \code{alpha} input value.
}
\details{
There are two unknown quantities in the DRpower model: the
  prevalence and the ICC. This function integrates out the prevalence over a
  prior distribution (we 'marginalise' out the prevalence) to give a
  distribution just in terms of the ICC. Then it returns the credible
  intervals of this distribution at a specified significance level.
}
\examples{
# define three clusters with different number of observed positive counts.
# Try the default 95\% CrI and a more stringent significance level
get_credible_ICC(pos_samples = c(2, 5, 4), total_samples = 10)
get_credible_ICC(pos_samples = c(2, 5, 4), total_samples = 10, alpha = 0.01)
}
