% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes.R
\name{get_credible_prevalence}
\alias{get_credible_prevalence}
\title{Get credible intervals for the prevalence over clusters}
\usage{
get_credible_prevalence(
  pos_samples,
  total_samples,
  alpha = 0.05,
  prior_prev_shape1 = 1,
  prior_prev_shape2 = 1,
  prior_ICC_shape1 = 1,
  prior_ICC_shape2 = 1,
  debug_on = FALSE
)
}
\arguments{
\item{pos_samples}{number of "positive" samples per cluster.}

\item{total_samples}{total sample size per cluster.}

\item{alpha}{the significance level of the credible interval - for example,
use \code{alpha = 0.05} for a 95\% interval.}

\item{prior_prev_shape1, prior_prev_shape2, prior_ICC_shape1, prior_ICC_shape2}{parameters that dictate the shape of the priors on prevalence and the ICC.
Increasing the first shape parameter (e.g. \code{prior_p_shape1}) pushes
the distribution to the right, increasing the second shape parameter (e.g.
\code{prior_p_shape2}) pushes the distribution to the left. Increasing both
shape parameters squeezes the distribution and makes it narrower.}

\item{debug_on}{if \code{TRUE} then produces a plot of the distribution of
the prevalence for use in debugging (for advanced users only). Open circles
are the Gaussian quadrature estimates, the red solid line is the
brute-force approach via trapezoidal rule, and green closed circles and
lines are the nodes and quadratic interpolations used in Simpson's rule -
these are used to produce final credible interval estimates.}
}
\description{
Produces lower and upper credible intervals on the prevalence
  from clustered counts. By default these are 95\% credible intervals,
  although the user can choose any significance level by altering the
  \code{alpha} input value.
}
\details{
There are two unknown quantities in the DRpower model: the
  prevalence and the ICC. This function integrates out the ICC over a prior
  distribution (we 'marginalise' out the ICC) to give a distribution just in
  terms of the prevalence. Then it returns the credible intervals of this
  distribution at a specified significance level.
}
\examples{
# define three clusters with different number of observed positive counts.
# Try the default 95\% CrI and a more stringent significance level
get_credible_prevalence(pos_samples = c(2, 5, 4), total_samples = 10)
get_credible_prevalence(pos_samples = c(2, 5, 4), total_samples = 10, alpha = 0.01)
}
