% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_credible_prevalence}
\alias{get_credible_prevalence}
\title{Get credible intervals for prevalence}
\usage{
get_credible_prevalence(
  pos_samples,
  total_samples,
  alpha = 0.05,
  prior_prev_shape1 = 1,
  prior_prev_shape2 = 1,
  prior_ICC_shape1 = 1,
  prior_ICC_shape2 = 1
)
}
\arguments{
\item{pos_samples}{number of "positive" samples per cluster.}

\item{total_samples}{total sample size per cluster.}

\item{alpha}{the significance level of the credible interval - for example,
use \code{alpha = 0.05} for a 95\% interval.}

\item{prior_prev_shape1, prior_prev_shape2, prior_ICC_shape1, prior_ICC_shape2}{parameters that dictate the shape of the priors on prevalence and the ICC.
Increasing the first shape parameter (e.g. \code{prior_p_shape1}) pushes
the distribution to the right, increasing the second shape parameter (e.g.
\code{prior_p_shape2}) pushes the distribution to the left. Increasing both
shape parameters squeezes the distribution and makes it narrower.}
}
\description{
Get lower and upper credible intervals on the prevalence over
  all clusters.
}
\examples{
# TODO
print("foo")
}
