% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_credible_prevalence}
\alias{get_credible_prevalence}
\title{Get credible intervals for prevalence}
\usage{
get_credible_prevalence(k, m, alpha = 0.05, p_nodes = 20, rho_nodes = 20)
}
\arguments{
\item{k}{counts of positive samples.}

\item{m}{sample size per cluster.}

\item{alpha}{the significance level of the credible interval - for sample,
use \code{alpha = 0.05} for 95}

\item{p_nodes, rho_nodes}{number of nodes to use in the Gaussian quadrature
approximate integral method.}
}
\description{
Get lower and upper credible intervals on the prevalence over
  all clusters.
}
\examples{
# TODO
print("foo")
}
