% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_quadrature.R
\name{get_HDI}
\alias{get_HDI}
\title{Get High Density Interval from adaptive quadrature object}
\usage{
get_HDI(df_norm, alpha = 0.05)
}
\arguments{
\item{df_norm}{a normalised quadrature data.frame as produced by
\code{normalise_quadrature()}.}

\item{alpha}{the significance level of the interval. Equivalently, one minus
the area under the curve of the interval.}
}
\description{
The HDI for a given alpha is defined as the narrowest interval
  in which the area under curve equals 1 - alpha. This can be obtained by
  searching down in the y-axis until the integrated area equals 1 - alpha.
  The HDI method is one method of calculating credible intervals.
}
