% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{get_margin_Bayesian}
\alias{get_margin_Bayesian}
\title{Margin of error calculations using the Bayesian DRpower model when
  estimating prevalence from a clustered survey}
\usage{
get_margin_Bayesian(
  N,
  prevalence = 0.2,
  ICC = 0.05,
  alpha = 0.05,
  prior_prev_shape1 = 1,
  prior_prev_shape2 = 1,
  prior_ICC_shape1 = 1,
  prior_ICC_shape2 = 9,
  CrI_type = "HDI",
  n_intervals = 20,
  round_digits = 2,
  reps = 100,
  use_cpp = TRUE,
  return_full = FALSE,
  silent = FALSE
)
}
\arguments{
\item{N}{vector of the number of samples obtained from each cluster.}

\item{prevalence}{assumed true prevalence of pfhrp2/3 deletions as a
proportion between 0 and 1.}

\item{ICC}{assumed true intra-cluster correlation (ICC) as a value between 0
and 1.}

\item{alpha}{the significance level of the credible interval - for example,
use \code{alpha = 0.05} for a 95\% interval. See also \code{CrI_type}
argument for how this is calculated.}

\item{prior_prev_shape1, prior_prev_shape2, prior_ICC_shape1, prior_ICC_shape2}{parameters that dictate the shape of the Beta priors on prevalence and the
ICC. See the \href{https://en.wikipedia.org/wiki/Beta_distribution}{Wikipedia page on the Beta
distribution} for more detail. The default values of these parameters were chosen
based on an
\href{https://mrc-ide.github.io/DRpower/articles/historical_analysis.html}{analysis
of historical pfhrp2/3 studies}, although this does not guarantee that they
will be suitable in all settings.}

\item{CrI_type}{which method to use when computing credible intervals.
Options are "ETI" (equal-tailed interval) or "HDI" (high-density interval).
The ETI searches a distance \code{alpha/2} from either side of the [0,1]
interval. The HDI method returns the narrowest interval that subtends a
proportion \code{1-alpha} of the distribution. The HDI method is used by
default as it guarantees that the MAP estimate is within the credible
interval, which is not always the case for the ETI.}

\item{n_intervals}{the number of intervals used in the adaptive quadrature
method. Increasing this value gives a more accurate representation of the
true posterior, but comes at the cost of reduced speed.}

\item{round_digits}{the number of digits after the decimal point that are
used when reporting estimates. This is to simplify results and to avoid
giving the false impression of extreme precision.}

\item{reps}{number of times to repeat simulation per parameter combination.}

\item{use_cpp}{if \code{TRUE} (the default) then use an Rcpp implementation
of the adaptive quadrature approach that is much faster than the base R
method.}

\item{return_full}{if \code{TRUE} then return the complete distribution of
lower and upper CrI limits in a data.frame. If \code{FALSE} (the default)
return a summary including the mean and 95\% CI of these limits.}

\item{silent}{if \code{TRUE} then suppress all console output.}
}
\value{
If \code{return_full = FALSE} (the default) returns a data.frame
  where the lower MOE is in the first row, and the upper MOE is in the second
  row. The first column gives the point estimate, and the subsequent columns
  give the 95\% CI on this estimate. If \code{return_full = TRUE} then
  returns a complete data.frame of all lower and upper MOE realisations over
  simulations.
}
\description{
As well as comparing against a threshold, the function
  \code{get_prevalence()} can be used to estimate a Bayesian credible
  interval (CrI) on the prevalence. We can estimate the margin of error (MOE)
  of this approach by simulation; repeatedly simulating datasets from known
  parameters, running \code{get_prevalence()} and measuring the MOE. Advantages of
  this method are that it accounts for uncertainty in the ICC, accounts for
  uncertainty in prevalence estimation by using simulation, and allows for
  incorporation of prior information.
}
\details{
Estimates MOE using the following approach:
\enumerate{
 \item Simulate data via the function \code{rbbinom_reparam()} using known
 values (e.g. a known "true" prevalence and intra-cluster correlation).
 \item Analyse data using \code{get_prevalence()} to determine the
 upper and lower limits of the credible interval.
 \item Repeat steps 1-2 many times to obtain the distribution of upper and
 lower limits Return either the full distribution, or the mean of the
 distribution along with upper and lower 95\% CIs.
}
Note that we have not implemented a function that returns the sample size
needed to achieve a given MOE under the Bayesian model because this would
require repeated simulation over different values of \code{N}, which is
computationally costly. The appropriate value can be established manually if
needed by running \code{get_margin_Bayesian()} for different sample sizes.
}
\examples{
get_margin_Bayesian(N = c(120, 90, 150), prevalence = 0.15, ICC = 0.01 , reps = 1e2)

}
