% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_power}
\alias{get_power}
\title{Estimate power via simulation}
\usage{
get_power(
  N,
  prevalence = 0.1,
  ICC = 0.1,
  prev_thresh = 0.05,
  rejection_threshold = 0.95,
  prior_prev_shape1 = 1,
  prior_prev_shape2 = 1,
  prior_ICC_shape1 = 1,
  prior_ICC_shape2 = 9,
  n_intervals = 20,
  reps = 100
)
}
\arguments{
\item{N}{vector giving the number of samples obtained from each cluster.}

\item{prevalence}{assumed true prevalence of pfhrp2 deletions. Input as
proportion between 0 and 1.}

\item{ICC}{assumed true intra-cluster correlation (ICC), between 0 and 1.}

\item{prev_thresh}{the threshold prevalence that we are testing against.}

\item{rejection_threshold}{the posterior probability of being above the
prevalence threshold needs to be greater than \code{rejection_threshold} in
order to reject the null hypothesis.}

\item{prior_prev_shape1, prior_prev_shape2, prior_ICC_shape1, prior_ICC_shape2}{parameters that dictate the shape of the Beta priors on prevalence and the
ICC. Increasing the first shape parameter (e.g. \code{prior_prev_shape1})
pushes the distribution towards 1, increasing the second shape parameter
(e.g. \code{prior_prev_shape2}) pushes the distribution towards 0.
Increasing both shape parameters squeezes the distribution towards the
centre and therefore makes it narrower. The default values of these
parameters are based on an analysis of historical pfhrp2/3 studies.}

\item{n_intervals}{the number of intervals used in the adaptive quadrature
method. Increasing this value gives a more accurate representation of the
true posterior, but comes at the cost of reduced speed.}

\item{reps}{number of times to repeat simulation per parameter combination.}
}
\description{
For a given sample design, estimates power empirically via
  repeated simulation (see Details). Returns an estimate of the power, along
  with lower and upper bounds of this estimate.
}
\details{
Estimates power using the following approach:
\enumerate{
 \item Simulate repeatedly from the function \code{rbbinom_reparam()} using
 known values (e.g. a known "true" prevalence and intra-cluster correlation).
 \item Analyse data using \code{get_prevalence()} to determine the
 probability of being above \code{prev_thresh}.
 \item If this probability is above \code{rejection_threshold} then
 reject the null hypothesis, and encode this as a single correct conclusion.
 \item Count the number of simulations for which the correct conclusion is
 reached, relative to the total number of simulations. This gives an estimate
 of empirical power, along with upper and lower 95\% binomial CIs on the power
 via the method of Clopper and Pearson (1934).
}
}
\examples{
get_power(N = c(120, 90, 150), prevalence = 0.15, ICC = 0.1 , reps = 1e2)

}
\references{
Clopper, C.J. and Pearson, E.S., 1934. The use of confidence or fiducial
limits illustrated in the case of the binomial. Biometrika, 26, 404–413. doi:
10.2307/2331986.
}
