% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_quadrature.R
\name{normalise_quadrature}
\alias{normalise_quadrature}
\title{Normalise an adaptive quadrature data.frame}
\usage{
normalise_quadrature(df_quad)
}
\arguments{
\item{df_quad}{a data.frame output from the function
\code{adaptive_quadrature()}.}
}
\description{
Takes data.frame output from the function
  \code{adaptive_quadrature()} and normalises the y-values so the curve
  integrates to 1 via Simpson's rule. Also reorders rows, drops some columns,
  and appends the Simpson's rule coefficients of the quadratic Ax^2 + Bx + C
  for each interval.
}
